+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/repocutter/compare/v0.5.0...HEAD)
 ------------------------------------------------------------------------
 ### Fixed
+- Evaluate template variables in cookiecutter.json
 - Fix error raised if already on revision with `-b/--branch`
 
 [0.5.0](https://github.com/jshwi/repocutter/releases/tag/v0.5.0) - 2023-01-22